refactor(auth): extract form logic into lazy-loaded client components
Convert sign-in and sign-up pages from Client Components to Server
Components, extracting all Zod/RHF form logic into separate
`sign-in-form.tsx` and `sign-up-form.tsx` files loaded via dynamic().

This follows the Server Component + lazy-loaded Client Component pattern:
pages render the AuthCard shell on the server, then stream in the
interactive form once JS is ready.

Note: Turbopack co-bundles Zod with next/link in a shared chunk
(41ab505) — this is a build-tool-level decision outside app code control.
